requirements.txt
python-dotenv
gradio[mcp]
google-genai